+++ src/resources/constants.rs
@@ -1,5 +1,5 @@
 pub const TILE_SIZE: f32 = 16.0;
-pub const PLANET_RADIUS: isize = 4;
+pub const PLANET_RADIUS: i32 = 4;
 
 pub const MOVE_COST: u32 = 100;
 pub const TURN_COST: u32 = 0;
